[IMP] website: improve Google Fonts GDPR compliance with cookies bar
__Current behavior before commit:__
When the cookies bar is installed, Google fonts are still loaded from
Google servers by default, which may violate GDPR requirements.

See the PR adding the "Serve from Google"[1] option for more details.

__Description of the change:__
If the cookies bar is enabled, it likely means that the website needs to
be GDPR compliant. In this case the "Serve from Google" option is
disabled by default when adding a new Google Font, preventing it from
being served from Google servers.
The tooltip and setting help text are also updated to clarify GDPR
implications.

[1]: https://github.com/odoo/odoo/pull/101129

task-5111612

closes odoo/odoo#235155

X-original-commit: ffe8cdc1e36080aff9ef84cece09c69c72f6446d
Signed-off-by: Outagant Mehdi (mou) <mou@odoo.com>
Signed-off-by: Julien Launois (jula) <jula@odoo.com>